Remove 'run_external' internal command.
We've resorted to an internal `run_external` command to express a command that we
are interested in shelling out. To build this command at parse time, we've resorted
to a few techniques starting from the positionals passed and feeding it too to
the `run_external`.

An InternalCommand struct is used to denote a command's name and call information at parse time.
We can use directly this same struct to express commands that no signature is available but rather
than classify it as an internal command we can classify it as an external and still use the struct.

For this case InternalCommand has been renamed to `CommandSpecification` to avoid confusion.

This may aid when external signatures for (say completions) in the future and has less indirection
when working with commands not supported.