Add subcommand `str to-bytes`
It's currently not possible to convert a string containing a number into a byte
unit. The only way to create an instance of a "bytes" type today is with a
literal in nushell syntax. This commit adds the `str to-bytes` subcommand so
that bytes can be created from the outputs of programs producing normal strings,
like standard unix tools.

There is a limitation with this - it doesn't currently parse values like `10 MB`
or `10 MiB`, it can only look at the number itself. If the desire is there, more
flexible parsing can be added.