Use rustls for TLS handshakes (fetch/post --commands)
This will make it easier building Nu without dynamic linking
suprises with OpenSSL. It's possible to statically link OpenSSL
but it is not simple to do.

With rustls we may do well.

We still need to decide what to do with `bat` crate (we seem to
use it with the `git` feature (which requires `git2` that ends up
pulling in OpenSSL dependencies as well).