Add a unicode folder for the completion tests
This does not excercise a particular difficult path of the completer
code, but should primarily serve as a reminder and backstop, that we
have to support complex Unicode containing paths or symbols in the
completion logic.

When doing string manipulation around the completions, think about
adding Unicode tests that place a character boundary around string
splitting logic!