Simplify counting/splitting of leading dots
Replaces the regex solution with an iterator-driven solution.

Checking for an ASCII character in bytes is OK as UTF multibyte
characters do not contain ASCII. Using `str.bytes()` is also nice for two
reasons, it makes it explicit that the index will be coherent with the
byte index in `str.split_at()` and this doens't require the UTF-8 parsing
of the `Chars` iterator.

Still open is dealing with files that have multiple dots at the
beginning. They currently get ignored.