[Experiment] Reenable CI build cache for tests
Let's see if we can again use `cargo-cache` for the tests after #6389
reduced the number of test binaries to build that are quite large due as
they statically link copies of the same engine.
This might be one of the reasons why the tests on windows exceeded the
alotted disk space.